+++ src/test/apps/service-task/App/Code/FailServiceTask.cs
@@ -1,4 +1,5 @@
-﻿using System.Threading.Tasks;
+﻿using System;
+using System.Threading.Tasks;
 using Altinn.App.Core.Internal.Data;
 using Altinn.App.Core.Features.Process;
 using Altinn.App.Core.Models;
@@ -22,7 +23,11 @@ public class FailServiceTask : IServiceTask
     {
         Instance instance = context.InstanceDataMutator.Instance;
 
-        DataElement dataElement2 = instance.Data.Find(x => x.DataType == "Model2");
+        DataElement dataElement2 =
+            instance.Data.Find(x => x.DataType == "Model2")
+            ?? throw new InvalidOperationException(
+                "Expected a 'Model2' data element on the instance."
+            );
 
         var formDataModel2 = (Model2)
             await context.InstanceDataMutator.GetFormData(new DataElementIdentifier(dataElement2));